Change factura constructor

+++ src/javafactura/businessLogic/econSectors/Familia.java
@@ -17,7 +17,7 @@ public final class Familia extends EconSector implements Deductible {
     }
 
     @Override
-    public float deduction(float value, boolean interior, int aggregateSize, float coefEmpresa, float coefIndividual){
+    public float deduction(float value, boolean interior, int aggregateSize, double coefEmpresa, double coefIndividual){
         return 0;
     }
 }
